lpc1xxx support
