+++ smashlib/lpc1xxx.py
@@ -1,4 +1,4 @@
-"""LPC1114 FLASH MEMORY PROGRAMMING FIRMWARE"""
+"""LPC1769 FLASH MEMORY PROGRAMMING FIRMWARE"""
 import os
 import time
 import binascii
@@ -9,7 +9,7 @@ from .micro import micro_info
 from .micro import ProtoError, IspTimeoutError, IspChecksumError, IspProgError
 from .binfile import BinFile
 
-class LPC111x(object):
+class LPC1xxx(object):
     def __init__(self, micro, freq, serial):
         self.micro = micro
         self.freq = freq
@@ -56,7 +56,8 @@ class LPC111x(object):
         self.serial.write(cmd + b"\r\n")
         if expected:
             expected += b"\r\n"
-            cmd += b"\r\n"
+            cmdtrail = micro_info[self.micro]["cmdtrail"]
+            cmd += cmdtrail
         else:
             cmd += b"\r\n"
 
@@ -85,16 +86,39 @@ class LPC111x(object):
     def sync_baudrate(self):
         self.retry(8, self._sync_baudrate, ())
 
+    def _set_reset(self, val):
+        self.serial.set_dtr(val)
+
+    def _set_psen(self, val):
+        self.serial.set_rts(val)
+
     def reset(self, isp):
-        pass
+        """RESET the Microcontroller"""
+        if isp == 1:
+            self._set_psen(1)
+            time.sleep(0.1)
+            self._set_reset(1)
+            time.sleep(0.1)
+            self._set_reset(0)
+            time.sleep(0.1)
+        else:
+            self._set_psen(0)
+            time.sleep(0.1)
+            self._set_reset(1)
+            time.sleep(0.1)
+            self._set_reset(0)
+            time.sleep(0.1)
 
     def unlock_and_checkid(self):
         """To unlock erase, write commands and
         to check the device we are going to flash.
         U 23130 --> Used to unlock flash write and erase commands
         J --> is used to read the part identification number """
+
+        devid = micro_info[self.micro]["devid"]
+
         self.write_and_expect(b"U 23130", b"0")
-        self.write_and_expect(b"J", b"0\r\n624955435")
+        self.write_and_expect(b"J", b"0\r\n" + devid.encode("ascii"))
 
     def set_osc_freq(self):
         """To Set the CCLK frequency,
@@ -102,6 +126,7 @@ class LPC111x(object):
         self.write_and_expect(b"12000", b"OK")
 
         self.unlock_and_checkid()
+
 
     def erase_block(self, block):
         """To Erase Selected Sectors in Flash"""
@@ -131,7 +156,9 @@ class LPC111x(object):
 
     def select_sector(self):
         """Select the Sectors for COPY command"""
-        sector_size = 4096
+        block_range = micro_info[self.micro]["block_range"]
+        block_info = block_range[self.start_sector]
+        sector_size = block_info[1] - block_info[0] + 1
 
         if self.sector_used == sector_size:
             self.start_sector += 1
@@ -256,12 +283,41 @@ common_sparams = {
 }
 
 micro_info.update({
+    "LPC1343" : {
+        "mfg": "NXP",
+        "block_range": ((0x0, 0xFFF), (0x1000, 0x1FFF), (0x2000, 0x2FFF), (0x3000, 0x3FFF),
+                        (0x4000, 0x4FFF), (0x5000, 0x5FFF), (0x6000, 0x6FFF), (0x7000, 0x7FFF)),
+        "sparams": common_sparams,
+        "class": LPC1xxx,
+        "security": ("CRP1", "CRP2", "CRP3"),
+        "cmdtrail": b"\r",
+        "devid": "1023410219"
+                },
     "LPC1114" : {
         "mfg": "NXP",
         "block_range": ((0x0, 0xFFF), (0x1000, 0x1FFF), (0x2000, 0x2FFF), (0x3000, 0x3FFF),
                         (0x4000, 0x4FFF), (0x5000, 0x5FFF), (0x6000, 0x6FFF), (0x7000, 0x7FFF)),
         "sparams": common_sparams,
-        "class": LPC111x,
-        "security": ("CRP1", "CRP2", "CRP3")
-        }
+        "class": LPC1xxx,
+        "security": ("CRP1", "CRP2", "CRP3"),
+        "cmdtrail": b"\r\n",
+        "devid": "624955435"
+                },
+    "LPC1769" : {
+        "mfg": "NXP",
+        "block_range": ((0x0, 0xFFF), (0x1000, 0x1FFF), (0x2000, 0x2FFF), (0x3000, 0x3FFF),
+                        (0x4000, 0x4FFF), (0x5000, 0x5FFF), (0x6000, 0x6FFF), (0x7000, 0x7FFF),
+                        (0x8000, 0x8FFF), (0x9000, 0x9FFF), (0xA000, 0xAFFF), (0xB000, 0xBFFF),
+                        (0xC000, 0xCFFF), (0xD000, 0xDFFF), (0xE000, 0xEFFF), (0xF000, 0xFFFF),
+                        (0x10000, 0x17FFF), (0x18000, 0x1FFFF), (0x20000, 0x27FFF),
+                        (0x28000, 0x2FFFF), (0x30000, 0x37FFF), (0x38000, 0x3FFFF),
+                        (0x40000, 0x47FFF), (0x48000, 0x4FFFF), (0x50000, 0x57FFF),
+                        (0x58000, 0x5FFFF), (0x60000, 0x67FFF), (0x68000, 0x6FFFF),
+                        (0x70000, 0x77FFF), (0x78000, 0x7FFFF)),
+        "sparams": common_sparams,
+        "class": LPC1xxx,
+        "security": ("CRP1", "CRP2", "CRP3"),
+        "cmdtrail": b"\r",
+        "devid": "638664503"
+                }
     })
